Problem: crate doesn't compile against latest pgx
Solution: upgrade to pgx 0.7.1

I've reduced the amount of coverage available for checked commands to
cover sub-transactions only until a better solution is found. The
problem right now is it it's hard to recover the parent. Previously
this was more feasible when SpiClient didn't need to be available
mutably.

This should solve cases it's being used for.